Moving reachability over from datastore to api category
