+++ AmplifyPlugins/DataStore/Podfile
@@ -9,7 +9,6 @@ pod 'SwiftLint'
 target 'AWSDataStoreCategoryPlugin' do
 	pod 'Amplify', :path => '../../'
   pod 'AWSPluginsCore', :path => '../../'
-  pod "ReachabilitySwift", "~> 5.0.0"
   pod "SQLite.swift", "~> 0.12.0"
 
   target "AWSDataStoreCategoryPluginTests" do
@@ -18,7 +17,6 @@ target 'AWSDataStoreCategoryPlugin' do
     pod "CwlPreconditionTesting", :git => "https://github.com/mattgallagher/CwlPreconditionTesting.git", :tag => "1.2.0"
     pod "CwlCatchException", :git => "https://github.com/mattgallagher/CwlCatchException.git", :tag => "1.2.0"
     pod 'AmplifyTestCommon', :path => '../../'
-    pod "ReachabilitySwift", "~> 5.0.0"
   end
 end
